refactor(dd_cascade): split the cascades and floatcascade into core / manipulators / iostream (#1334)
Final phase of the header layering: dd_cascade, td_cascade, qd_cascade and the
floatcascade substrate they share.

    internal/floatcascade/floatcascade.hpp   66,388 lines, 0 I/O-family (was 84,315 / 5)
    dd_cascade/core.hpp                      75,583 lines, 0 I/O-family (umbrella 87,005 / 5)
    td_cascade/core.hpp                      75,384 lines, 0 I/O-family (umbrella 87,245 / 5)
    qd_cascade/core.hpp                      75,469 lines, 0 I/O-family (umbrella 87,347 / 5)

The substrate was the real work. floatcascade.hpp defined operator<< as an IN-CLASS FRIEND
DEFINITION, which is what pinned <iostream> into every consumer -- all three cascade types
plus the floatcascade tests. It is now a friend declaration, so the core needs only
<iosfwd>, with the definition out-of-line in floatcascade/iostream.hpp; to_tuple() and
to_scientific() moved to floatcascade/manipulators.hpp.

The three cascade types had the identical in-class-friend pattern. Their to_string() is
public, so operator<< never needed friendship at all and became a plain free function.

Because that operator is a template, a clean -fsyntax-only proves nothing -- it is not
instantiated until used. It was instantiated and the output diffed against main:
byte-identical, including the fc.e[i] -> fc[i] substitution (operator[] is literally
`return e[i];`).

Division as for dd (#1426) and qd (#1427): to_string() and parse() stay in the CORE,
because assign(const std::string&) calls parse() and to_string() concatenates a
std::string without opening a stream. Diagnostics moved to fprintf, including the four in
floatcascade.hpp. manipulators.hpp and attributes.hpp include iostream.hpp for the reason
established in #1426: their builders format the type THROUGH operator<<. Not a cycle --
iostream.hpp includes only core.hpp.

td_cascade and qd_cascade carried the same color_print recursion that took down Linux CI
on #1427: they format each limb with color_print(number[i]) on a double, whose intended
target is the NON-TEMPLATE color_print(double) in native/ieee754_double.hpp, which
ieee754_core.hpp deliberately omits. Both now take native/ieee754.hpp in their TEXT layer.
Binding color_print<double> explicitly is NOT equivalent -- it selects the color_print<Real>
template and prints a different format -- so the include is the fix, not the call site.
Verified byte-identical to main through the umbrella AND with manipulators.hpp as the only
include. dd_cascade was never affected.

Closed 21 IWYU gaps across the four subtrees: <limits>, <cstdint>, <cstddef>, <cassert>,
<cmath>, and <string> in three exceptions.hpp plus td_cascade_fwd.hpp, which had the same
parse(const std::string&)-without-<string> defect as qd_fwd.hpp. Sweep clean at 0 gaps
across all 34 headers; every header already carried #pragma once.

Verified: all three core gates pass on gcc and clang, with #error tripwires on the stream
include guards and exact dyadic identities so the assertions do not depend on cascade
rounding; sweep clean; ASCII clean.

This completes Phase 2 of #1334.
