+++ include/sw/universal/number/dd_cascade/attributes.hpp
@@ -5,6 +5,12 @@
 // SPDX-License-Identifier: MIT
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
+#include <string>
+#include <sstream>
+#include <iomanip>
+#include <limits>       // std::numeric_limits
+#include <universal/number/dd_cascade/core.hpp>
+#include <universal/number/dd_cascade/iostream.hpp>   // the range reporters stream dd_cascade values through operator<<
 #include <cmath>
 
 namespace sw { namespace universal {
