+++ include/sw/universal/number/qd_cascade/qd_cascade.hpp
@@ -13,8 +13,6 @@
 
 ////////////////////////////////////////////////////////////////////////////////////////
 /// required std libraries
-#include <iostream>
-#include <iomanip>
 
 ////////////////////////////////////////////////////////////////////////////////////////
 ///  BEHAVIORAL COMPILATION SWITCHES
@@ -43,24 +41,27 @@
 
 ///////////////////////////////////////////////////////////////////////////////////////
 // bring in the trait functions
-#include <universal/traits/number_traits.hpp>
+/// INCLUDE FILES that make up the library
+///
+/// Layered per #1334:
+///     core.hpp          arithmetic, no streams
+///     manipulators.hpp  type_tag / to_binary / to_triple
+///     iostream.hpp      operator<< / operator>>
+///
+/// A translation unit that only computes can include core.hpp directly and skip the
+/// stream headers entirely. to_string() and parse() are in the CORE, not the text
+/// layers: assign(const std::string&) calls parse(), and to_string() concatenates a
+/// std::string without ever touching a stream.
+#include <universal/number/qd_cascade/core.hpp>
+#include <universal/traits/qd_cascade_traits.hpp>
+#include <universal/number/qd_cascade/manipulators.hpp>
+#include <universal/number/qd_cascade/iostream.hpp>
+#include <universal/number/qd_cascade/attributes.hpp>
+
+/// the report builders stay at the umbrella: they pull <sstream>/<iomanip> by design
 #include <universal/traits/arithmetic_traits.hpp>
 #include <universal/common/number_traits_reports.hpp>
 
-////////////////////////////////////////////////////////////////////////////////////////
-/// INCLUDE FILES that make up the library
-#include <universal/number/qd_cascade/exceptions.hpp>
-#include <universal/number/qd_cascade/qd_cascade_fwd.hpp>
-#include <universal/number/qd_cascade/qd_cascade_impl.hpp>
-#include <universal/number/qd_cascade/numeric_limits.hpp>
-#include <universal/traits/qd_cascade_traits.hpp>
-
-// useful functions to work with quad-doubles
-#include <universal/number/qd_cascade/manipulators.hpp>
-#include <universal/number/qd_cascade/attributes.hpp>
-
-///////////////////////////////////////////////////////////////////////////////////////
-/// mathematical constants library
 #include <universal/number/qd_cascade/math/constants/qd_cascade_constants.hpp>
 
 ///////////////////////////////////////////////////////////////////////////////////////
