+++ include/sw/universal/number/td_cascade/exceptions.hpp
@@ -5,6 +5,7 @@
 // SPDX-License-Identifier: MIT
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
+#include <string>       // std::string, the exception message type
 #include <universal/common/exceptions.hpp>
 
 namespace sw { namespace universal {
