+++ include/sw/universal/number/td_cascade/td_cascade_fwd.hpp
@@ -5,6 +5,7 @@
 // SPDX-License-Identifier: MIT
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
+#include <string>       // std::string, in the parse() declaration below
 
 namespace sw { namespace universal {
